get works

+++ source/socket/response/Delete.cpp
@@ -1,22 +1,35 @@
 #include "Delete.hpp"
 
-Delete::Delete() {}
+Delete::Delete(Server & server, Request & request, int & client_socket) : AMethods(server, request, DELETE, client_socket) {}
 
-Delete::Delete(Delete const & src) {
-    *this = src;
-}
+Delete::Delete(Delete const & src) : AMethods(src._server, src._request, DELETE, src._client_socket) {}
 
 Delete::~Delete() {}
 
-Delete & Delete::operator=(Delete const & rhs) {
+Delete & Delete::operator=(Delete const & rhs)
+{
+	if (this != &rhs)
+	{
+        _method = rhs._method;
+        _location = rhs._location;
+        _path_file = rhs._path_file;
+		_server = rhs._server;
+		_request = rhs._request;
+		_client_socket = rhs._client_socket;
+	}
     return *this;
 }
 
-std::string Delete::execution(Server const & server, Request & request, int & client_socket)
+void Delete::execute(void)
 {
-	if (file_exist(this->_path_file))
-		remove_file(this->_path_file.c_str());
-	else
-		request.set_error(std::make_pair(404, "Not Found"));
-    return "";
+	if (this->search_location() == EXIT_SUCCESS)
+	{
+		if (this->is_method_allowed() == true)
+		{
+			if (file_exist(this->_path_file))
+				remove_file(this->_path_file.c_str());
+			else
+				_request.set_error(std::make_pair(404, "Not Found"));
+		}
+	}
 }
